backend/requirements.txt
fastapi
uvicorn[standard]
yfinance
pandas
google-genai
python-dotenv
